fix(clippy): resolve all clippy -D warnings for release CI
Fix 52 clippy warnings across 7 files: replace .max().min() chains with
.clamp(), add #[allow(clippy::too_many_arguments)] to 4 functions, merge
identical if-else branches in get_estimated_expense_ratio, and fix loop
variable indexing in compute_diversification_score.

+++ src-tauri/src/modules/data_provider/multi_source_provider.rs
@@ -178,7 +178,7 @@ impl MultiSourceProvider {
             if total == 0 {
                 return 100;
             }
-            (successes * 100 / total) as u32
+            successes * 100 / total
         } else {
             100 // New provider, assume healthy
         }
